// src/app/model/userList.model.ts
export interface UserList {
  user_id: string;
  first_name: string;
  last_name: string;
  email: string;
  password: string;
  profile_image: string;
  profile_image_thumb: string;
  token: string;
  dob: string;
  mobile_number: string;
  home_number: string;
  emergency_number: string;
  kin_name: string;
  kin_number: string;
  address_line_1: string;
  address_line_2: string;
  notification_count: string;
  post_code: string;
  city: string;
  status: string;
  created_at: string;
  employement_type: string;
  unique_id : string
}